Add ability to declare certificate pins in application manifest
Certificate pins specified by the application manifest in <cert-pin> entries
are added to the certificate pin list alongside the existing system-wide
pins in /data/misc/keychain/pins. This commit adds this capability
to the NetworkSecurityPolicy class.

This change goes along with corresponding changes to frameworks/base
and external/conscrypt.

Currently, application developers must override the platform's
TrustManager with a custom TrustManager in order to implement
certificate pinning. Overriding the TrustManager is a common
source of vulnerabilities as described by Tendulkar and Enck [1],
who suggest moving portions of TLS configuration such as
certificate pins out of application code and into the application
manifest. (Other references to common TLS vulnerabilities
in Android application include [2] [3] and [4].)
This change builds upon the system-wide certificate pinning
support that is now in Android to alleviate the need for
application developers to implement a custom TrustManager
for application-level certificate pinning.

[1] Tendulkar and Enck. An Application Package Configuration Approach to
Mitigating Android SSL Vulnerabilities. MoST 2014.
[2] Fahl, et al. Why Eve and Mallory Love Android: An Analysis of Android
SSL (In)Security. CCS '12.
[3] Sounthiraraj, et al. Large Scale, Automated Detection of SSL/TLS
Man-in-the-Middle Vulnerabilities in Android Apps. NDSS '14.
[4] FireEye. SSL Vulnerabilities: Who listens when Android applications talk?
https://www.fireeye.com/blog/threat-research/2014/08/ssl-vulnerabilities-who-listens-when-android-applications-talk.html

Change-Id: I63e5541bb2bc8a1bd353e0b01ace3666f55a82d4